+++ dukan_presentation/src/commonMain/kotlin/net/thechance/mena/dukan/presentation/viewModel/mainScreen/MainInteractionListener.kt
@@ -1,4 +1,4 @@
-package net.thechance.mena.dukan.presentation.screen.home
+package net.thechance.mena.dukan.presentation.viewModel.mainScreen
 
 interface MainInteractionListener {
     fun onDukanButtonClicked()
